Some int->size_t stuff
Using signed types for sizes is so stupid.

+++ linuxdoom-1.10/i_system.h
@@ -24,6 +24,8 @@
 #ifndef __I_SYSTEM__
 #define __I_SYSTEM__
 
+#include <stddef.h>
+
 #include "d_ticcmd.h"
 #include "d_event.h"
 
@@ -38,7 +40,7 @@ void I_Init (void);
 // Called by startup code
 // to get the ammount of memory to malloc
 // for the zone management.
-byte*	I_ZoneBase (int *size);
+byte*	I_ZoneBase (size_t *size);
 
 
 // Called by D_DoomLoop,
@@ -82,7 +84,7 @@ void I_Quit (void);
 
 // Allocates from low memory under dos,
 // just mallocs under unix
-byte* I_AllocLow (int length);
+byte* I_AllocLow (size_t length);
 
 void I_Tactile (int on, int off, int total);
